# lib/api-zod/src/generated/types/cashbackBalance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CashbackBalanceCurrency } from './cashbackBalanceCurrency';

export interface CashbackBalance {
  currency: CashbackBalanceCurrency;
  pending: number;
  available: number;
}